# electron/api/dist/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"GenericResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"userType": {
				"properties": {
					"balances": {
						"items": {},
						"type": "array"
					},
					"accounts": {
						"items": {
							"properties": {
								"caip": {
									"type": "string"
								},
								"pubkey": {}
							},
							"required": [
								"caip",
								"pubkey"
							],
							"type": "object"
						},
						"type": "array"
					},
					"online": {
						"type": "boolean"
					}
				},
				"required": [
					"balances",
					"accounts",
					"online"
				],
				"type": "object"
			},
			"PublicKey": {
				"properties": {
					"xpub": {
						"type": "string"
					}
				},
				"required": [
					"xpub"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BIP32Path": {
				"items": {
					"type": "number",
					"format": "double"
				},
				"type": "array"
			},
			"BTCInputScriptType": {
				"enum": [
					"cashaddr",
					"bech32",
					"p2pkh",
					"p2sh",
					"external",
					"p2wpkh",
					"p2sh-p2wpkh"
				],
				"type": "string"
			},
			"Coin": {
				"type": "string"
			},
			"GetPublicKey": {
				"properties": {
					"addressNList": {
						"$ref": "#/components/schemas/BIP32Path"
					},
					"showDisplay": {
						"type": "boolean"
					},
					"scriptType": {
						"$ref": "#/components/schemas/BTCInputScriptType"
					},
					"curve": {
						"type": "string"
					},
					"coin": {
						"$ref": "#/components/schemas/Coin"
					}
				},
				"required": [
					"addressNList",
					"curve",
					"coin"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BTCGetAddress": {
				"properties": {
					"showDisplay": {
						"type": "boolean"
					},
					"scriptType": {
						"$ref": "#/components/schemas/BTCInputScriptType"
					},
					"addressNList": {
						"$ref": "#/components/schemas/BIP32Path"
					},
					"coin": {
						"$ref": "#/components/schemas/Coin"
					}
				},
				"required": [
					"addressNList",
					"coin"
				],
				"type": "object"
			},
			"ETHGetAddress": {
				"properties": {
					"addressNList": {
						"$ref": "#/components/schemas/BIP32Path"
					},
					"showDisplay": {
						"type": "boolean"
					}
				},
				"required": [
					"addressNList"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ThorchainGetAddress": {
				"properties": {
					"addressNList": {
						"$ref": "#/components/schemas/BIP32Path"
					},
					"showDisplay": {
						"type": "boolean"
					},
					"testnet": {
						"type": "boolean"
					}
				},
				"required": [
					"addressNList"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OsmosisGetAddress": {
				"properties": {
					"addressNList": {
						"$ref": "#/components/schemas/BIP32Path"
					},
					"showDisplay": {
						"type": "boolean"
					}
				},
				"required": [
					"addressNList"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BinanceGetAddress": {
				"properties": {
					"addressNList": {
						"$ref": "#/components/schemas/BIP32Path"
					},
					"showDisplay": {
						"type": "boolean"
					}
				},
				"required": [
					"addressNList"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CosmosGetAddress": {
				"properties": {
					"addressNList": {
						"$ref": "#/components/schemas/BIP32Path"
					},
					"showDisplay": {
						"type": "boolean"
					}
				},
				"required": [
					"addressNList"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BTCSignedTx": {
				"properties": {
					"signatures": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"serializedTx": {
						"type": "string",
						"description": "hex string representation of the raw, signed transaction"
					}
				},
				"required": [
					"signatures",
					"serializedTx"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Thorchain.Msg": {
				"properties": {
					"type": {
						"type": "string"
					},
					"value": {}
				},
				"required": [
					"type",
					"value"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Coins": {
				"items": {
					"$ref": "#/components/schemas/Coin"
				},
				"type": "array"
			},
			"Thorchain.StdFee": {
				"properties": {
					"amount": {
						"$ref": "#/components/schemas/Coins"
					},
					"gas": {
						"type": "string"
					}
				},
				"required": [
					"amount",
					"gas"
				],
				"type": "object",
				"additionalProperties": false
			},
			"crypto.PubKey": {
				"properties": {
					"type": {
						"type": "string"
					},
					"value": {
						"type": "string"
					}
				},
				"required": [
					"type",
					"value"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Thorchain.StdSignature": {
				"properties": {
					"pub_key": {
						"$ref": "#/components/schemas/crypto.PubKey"
					},
					"signature": {
						"type": "string"
					}
				},
				"required": [
					"signature"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ThorchainTx": {
				"properties": {
					"msg": {
						"items": {
							"$ref": "#/components/schemas/Thorchain.Msg"
						},
						"type": "array"
					},
					"fee": {
						"$ref": "#/components/schemas/Thorchain.StdFee"
					},
					"signatures": {
						"items": {
							"$ref": "#/components/schemas/Thorchain.StdSignature"
						},
						"type": "array",
						"nullable": true
					},
					"memo": {
						"type": "string"
					}
				},
				"required": [
					"msg",
					"fee",
					"signatures",
					"memo"
				],
				"type": "object",
				"additionalProperties": false
			},
			"StdFee": {
				"properties": {
					"amount": {
						"$ref": "#/components/schemas/Coins"
					},
					"gas": {
						"type": "string"
					}
				},
				"required": [
					"amount",
					"gas"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Msg": {
				"properties": {
					"type": {
						"type": "string"
					},
					"value": {}
				},
				"required": [
					"type",
					"value"
				],
				"type": "object",
				"additionalProperties": false
			},
			"StdSignature": {
				"properties": {
					"pub_key": {
						"$ref": "#/components/schemas/crypto.PubKey"
					},
					"signature": {
						"type": "string"
					}
				},
				"required": [
					"signature"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Thorchain.StdTx": {
				"properties": {
					"fee": {
						"$ref": "#/components/schemas/StdFee"
					},
					"memo": {
						"type": "string"
					},
					"msg": {
						"items": {
							"$ref": "#/components/schemas/Msg"
						},
						"type": "array"
					},
					"signatures": {
						"items": {
							"$ref": "#/components/schemas/StdSignature"
						},
						"type": "array",
						"nullable": true
					}
				},
				"required": [
					"fee",
					"memo",
					"msg",
					"signatures"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ThorchainSignTx": {
				"properties": {
					"addressNList": {
						"$ref": "#/components/schemas/BIP32Path"
					},
					"tx": {
						"$ref": "#/components/schemas/Thorchain.StdTx"
					},
					"sequence": {
						"type": "string"
					},
					"account_number": {
						"type": "string"
					},
					"chain_id": {
						"type": "string"
					},
					"fee": {
						"type": "number",
						"format": "double"
					},
					"testnet": {
						"type": "boolean"
					}
				},
				"required": [
					"addressNList",
					"tx",
					"sequence",
					"account_number",
					"chain_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CosmosSignedTx": {
				"properties": {
					"serialized": {
						"type": "string"
					},
					"body": {
						"type": "string"
					},
					"authInfoBytes": {
						"type": "string"
					},
					"signatures": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"serialized",
					"body",
					"authInfoBytes",
					"signatures"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Cosmos.StdTx": {
				"properties": {
					"msg": {
						"items": {
							"$ref": "#/components/schemas/Msg"
						},
						"type": "array"
					},
					"fee": {
						"$ref": "#/components/schemas/StdFee"
					},
					"signatures": {
						"items": {
							"$ref": "#/components/schemas/StdSignature"
						},
						"type": "array",
						"nullable": true
					},
					"memo": {
						"type": "string"
					}
				},
				"required": [
					"msg",
					"fee",
					"signatures",
					"memo"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CosmosSignTx": {
				"properties": {
					"addressNList": {
						"$ref": "#/components/schemas/BIP32Path"
					},
					"tx": {
						"$ref": "#/components/schemas/Cosmos.StdTx"
					},
					"chain_id": {
						"type": "string"
					},
					"account_number": {
						"type": "string"
					},
					"sequence": {
						"type": "string"
					},
					"fee": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"addressNList",
					"tx",
					"chain_id",
					"account_number",
					"sequence"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ETHSignedTx": {
				"properties": {
					"v": {
						"type": "number",
						"format": "double",
						"description": "uint32"
					},
					"r": {
						"type": "string",
						"description": "big-endian hex, prefixed with '0x'"
					},
					"s": {
						"type": "string",
						"description": "big-endian hex, prefixed with '0x'"
					},
					"serialized": {
						"type": "string",
						"description": "big-endian hex, prefixed with '0x'"
					}
				},
				"required": [
					"v",
					"r",
					"s",
					"serialized"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ExchangeType": {
				"properties": {
					"signedExchangeResponse": {
						"type": "string",
						"description": "`SignedExchangeResponse` from the `/sendamountProto2` ShapeShift endpoint, base64 encoded"
					},
					"withdrawalCoinName": {
						"type": "string"
					},
					"withdrawalAddressNList": {
						"$ref": "#/components/schemas/BIP32Path"
					},
					"withdrawalScriptType": {
						"$ref": "#/components/schemas/BTCInputScriptType"
					},
					"returnAddressNList": {
						"$ref": "#/components/schemas/BIP32Path"
					},
					"returnScriptType": {
						"$ref": "#/components/schemas/BTCInputScriptType"
					}
				},
				"required": [
					"signedExchangeResponse",
					"withdrawalCoinName",
					"withdrawalAddressNList",
					"returnAddressNList"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ETHSignTx": {
				"properties": {
					"addressNList": {
						"$ref": "#/components/schemas/BIP32Path",
						"description": "bip32 path to sign the transaction from"
					},
					"nonce": {
						"type": "string",
						"description": "big-endian hex, prefixed with '0x'"
					},
					"gasPrice": {
						"type": "string",
						"description": "big-endian hex, prefixed with '0x'"
					},
					"gasLimit": {
						"type": "string",
						"description": "big-endian hex, prefixed with '0x'"
					},
					"maxFeePerGas": {
						"type": "string",
						"description": "EIP-1559 - The maximum total fee per gas the sender is willing to pay. <=256 bit unsigned big endian (in wei)"
					},
					"maxPriorityFeePerGas": {
						"type": "string",
						"description": "EIP-1559 - Maximum fee per gas the sender is willing to pay to miners. <=256 bit unsigned big endian (in wei)"
					},
					"to": {
						"type": "string",
						"description": "address, with '0x' prefix"
					},
					"toAddressNList": {
						"$ref": "#/components/schemas/BIP32Path",
						"description": "bip32 path for destination (device must `ethSupportsSecureTransfer()`)"
					},
					"value": {
						"type": "string",
						"description": "big-endian hex, prefixed with '0x'"
					},
					"data": {
						"type": "string",
						"description": "prefixed with '0x'"
					},
					"chainId": {
						"type": "number",
						"format": "double",
						"description": "mainnet: 1, ropsten: 3, kovan: 42"
					},
					"exchangeType": {
						"$ref": "#/components/schemas/ExchangeType",
						"description": "Device must `ethSupportsNativeShapeShift()`"
					}
				},
				"required": [
					"addressNList",
					"nonce",
					"gasLimit",
					"to",
					"value",
					"data",
					"chainId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SignedTx": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"status": {
						"type": "string"
					},
					"signedTx": {}
				},
				"required": [
					"success",
					"status",
					"signedTx"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Read": {
				"properties": {
					"data": {
						"type": "string"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Error": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"reason": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"reason"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Write": {
				"properties": {
					"output": {
						"type": "string"
					}
				},
				"required": [
					"output"
				],
				"type": "object",
				"additionalProperties": false
			},
			"WriteBody": {
				"properties": {
					"data": {}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Status": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"status": {
						"type": "string"
					},
					"state": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"status",
					"state"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Record_string.unknown_": {
				"properties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"PairResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"reason": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"reason"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PairBody": {
				"properties": {
					"serviceName": {
						"type": "string"
					},
					"serviceImageUrl": {
						"type": "string"
					}
				},
				"required": [
					"serviceName",
					"serviceImageUrl"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "keepkey-desktop",
		"version": "0.1.33",
		"contact": {
			"name": "Highlander ",
			"email": "bithighlander@gmail.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/auth/verify": {
			"get": {
				"operationId": "VerifyAuth",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GenericResponse"
								}
							}
						}
					},
					"401": {
						"description": "Please provice a valid serviceKey"
					}
				},
				"tags": [
					"Secured Endpoints"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": []
			}
		},
		"/user": {
			"get": {
				"operationId": "User",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/userType"
								}
							}
						}
					},
					"401": {
						"description": "Please provice a valid serviceKey"
					}
				},
				"tags": [
					"Secured Endpoints"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": []
			}
		},
		"/getPublicKeys": {
			"post": {
				"operationId": "GetPublicKeys",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"allOf": [
											{
												"$ref": "#/components/schemas/PublicKey"
											}
										],
										"nullable": true
									},
									"type": "array"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error"
					}
				},
				"tags": [
					"Secured Endpoints"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/GetPublicKey"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
		"/btcGetAddress": {
			"post": {
				"operationId": "BtcGetAddress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error"
					}
				},
				"tags": [
					"Secured Endpoints"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BTCGetAddress"
							}
						}
					}
				}
			}
		},
		"/ethGetAddress": {
			"post": {
				"operationId": "EthGetAddress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error"
					}
				},
				"tags": [
					"Secured Endpoints"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ETHGetAddress"
							}
						}
					}
				}
			}
		},
		"/thorchainGetAddress": {
			"post": {
				"operationId": "ThorchainGetAddress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string",
									"nullable": true
								}
							}
						}
					},
					"500": {
						"description": "Internal server error"
					}
				},
				"tags": [
					"Secured Endpoints"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ThorchainGetAddress"
							}
						}
					}
				}
			}
		},
		"/osmosisGetAddress": {
			"post": {
				"operationId": "OsmosisGetAddress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"500": {
						"description": "Internal server error"
					}
				},
				"tags": [
					"Secured Endpoints"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/OsmosisGetAddress"
							}
						}
					}
				}
			}
		},
		"/binanceGetAddress": {
			"post": {
				"operationId": "BinanceGetAddress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error"
					}
				},
				"tags": [
					"Secured Endpoints"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BinanceGetAddress"
							}
						}
					}
				}
			}
		},
		"/cosmosGetAddress": {
			"post": {
				"operationId": "CosmosGetAddress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error"
					}
				},
				"tags": [
					"Secured Endpoints"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CosmosGetAddress"
							}
						}
					}
				}
			}
		},
		"/btcSignTx": {
			"post": {
				"operationId": "BtcSignTx",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BTCSignedTx"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error"
					}
				},
				"tags": [
					"Secured Endpoints"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
		"/thorchainSignTx": {
			"post": {
				"operationId": "ThorchainSignTx",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ThorchainTx"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error"
					}
				},
				"tags": [
					"Secured Endpoints"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ThorchainSignTx"
							}
						}
					}
				}
			}
		},
		"/cosmosSignTx": {
			"post": {
				"operationId": "CosmosSignTx",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CosmosSignedTx"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error"
					}
				},
				"tags": [
					"Secured Endpoints"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CosmosSignTx"
							}
						}
					}
				}
			}
		},
		"/osmosisSignTx": {
			"post": {
				"operationId": "OsmosisSignTx",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"500": {
						"description": "Internal server error"
					}
				},
				"tags": [
					"Secured Endpoints"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
		"/ethSignTx": {
			"post": {
				"operationId": "EthSignTx",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ETHSignedTx"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error"
					}
				},
				"tags": [
					"Secured Endpoints"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ETHSignTx"
							}
						}
					}
				}
			}
		},
		"/sign": {
			"post": {
				"operationId": "SignTransaction",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SignedTx"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error"
					}
				},
				"tags": [
					"Secured Endpoints"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
		"/exchange/device": {
			"get": {
				"operationId": "ReadDevice",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/Read"
										},
										{
											"$ref": "#/components/schemas/Error"
										}
									]
								}
							}
						}
					},
					"500": {
						"description": "Unable to communicate with device"
					}
				},
				"tags": [
					"Raw KeepKey Device I/0 Endpoints"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "WriteDevice",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/Write"
										},
										{
											"$ref": "#/components/schemas/Error"
										}
									]
								}
							}
						}
					},
					"500": {
						"description": "Unable to communicate with device"
					}
				},
				"tags": [
					"Raw KeepKey Device I/0 Endpoints"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/WriteBody"
							}
						}
					}
				}
			}
		},
		"/status": {
			"get": {
				"operationId": "Status",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Status"
								}
							}
						}
					}
				},
				"tags": [
					"Client Endpoints"
				],
				"security": [],
				"parameters": []
			}
		},
		"/device": {
			"get": {
				"operationId": "Device",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Record_string.unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"Client Endpoints"
				],
				"security": [],
				"parameters": []
			}
		},
		"/pair": {
			"post": {
				"operationId": "Pair",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PairResponse"
								}
							}
						}
					},
					"500": {
						"description": ""
					}
				},
				"tags": [
					"Client Endpoints"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PairBody"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:1646/"
		}
	]
}